refactor tests to use inline data
